Manhattan room/appt dataframes; iqr column; groupby attempt
Added in dataframes for the Manahattan apartment and room categories to aid exploration of the outliers.
It looks like we will want both Manhattan and Brooklyn data to get over the 10k rows requirements for analysis.
Calculated iqr for each neighborhood independantly and there is a wide range for those values so that is likely
the primary groupping we'll want for analysis. When confirmed with the group, that probably means we will end
up recombining the Manhattan and Brooklyn values and just splitting by room vs appartment in each neighborhood.
Still trying to figure out how to independently apply IQR filtering to each neighborhood's data.